Tighten self-repair messages and split test store delegates (#35)
Pluralise successful self-repair messages for the single-iteration case
and update the concurrency test expectation to match the production text.

Move the workspace-store delegate implementation for `CapturingStore` into
its own module so the main delegation module stays below the file-size
limit. Also fix the self-repair helper wording in the developers guide.
